Fix for consistent firing of buttonup/down/touchstart/end/buttonchanged
This change delinks these event streams so that they can fire independenty of one another. I've also added a sufficiently complete number of tests for the new behaior to make sure all events that should be firing are, looking for negative events (ensuring some events don't fire when they should not) and that the values all come through as expected.

As part of the fix I've supplied some helpers that are now consistently used in tracked-controls.test.js for verifying individual events and streams of events in a more concise way to avoid duplicating a lot of code.

I've also made consistent the use of sinon's spy asserts, which were previously using the ".called" property which reaches into the abstraction further than it should. This also allows us to upgrade certain spy verification's to look for a specific number of calls (calledOnce) which is an overall improvement.

This change cascaded into the oculus-touch-controls.test.js and gearvr-controls.test.js as well. For the oculus-touch-controls, this is mostly assert fixups, but also does some additional deep object verification.

For the gearvr-controls.test.js this included some bug fixes (using wrong event names) when I was trying to make the test fail and could not, only to realize it wasn't sniffing for the right event names. This is in preparation for creation of the oculus-go-controls component and test suite.